Here is a paragraph about this commit.
Here is another paragraph.

This commit represents the semi-shitty physics that seems to work at 16x
speed. https://www.youtube.com/watch?v=R_xKoGWSGTg has a video of it
(please play at 16x for this to be bearable).

Some things to note: Density of 50k is a little too much, and 25k is
slightly too little. The config in this particular commit seemed to work
reasonably well.